Cleaned up LifetimeScope tests
Moved every class to own file.
